refactor middleware -> lib
